Notify users on asset checkout
Summary:
Send an email to users when an asset has been  assigned to them.

Send notification mail on asset checkout.

Send an email users indicating that an asset is now under their possession

Merge branch 'master' of https://github.com/livinggoods/lg-snipe-it

# Conflicts:
#	composer.lock
#	config/app.php

Send notification email on asset checkout

Added sendCheckOutAssignmentMail to send a notification to a user when an asset has been assigned to them.
Added text translations for the notification message.
Added an email template to format the notification that is sent to the user.

Test Plan:
  When an asset has been assigned, the user who has been assigned the asset will get a notification.
  - Select an asset, ideally one that has not been assigned.
  - Assign it to a user (who has a valid email address).
  - Check out the asset to the user.
  -- The user should receive an email. Check if they receive an email.

Reviewers: dkimaru

Reviewed By: dkimaru

Maniphest Tasks: T95

Differential Revision: https://dev.lg-apps.com/D17